Several changes: mostly comments and a few helper / util functions
Agent: comments for the methods and addition of utility (aciton filter) methods.
Action: comments and a few helper actions added.
ActionType: comments added to the class.
Entity: Comments and function added to get action types associated to the entity. Needs to pass a GameInfo reference, which is not ideal.
GameState: possibility of returning player action types by ID.