RTS and TBS ForwardModels unified for agents. See extended comments.
Functions in Agent.h for rolling the state forward (both RTS and TBS and opponent models).
Agent Parameters now uses preprocessFM for both portfolio on/off and RTS/TBS forward models. Also opponent model and heuristic accessors. New PortfolioRTSForwardModel.
Agents now use ForwardModel& rather than TBSForwardModel&
Tree Nodes compute action space considering which player can play at that game state. Each node holds an ownerID value that indicates who owns the tree that this node belongs to, and a playerID that indicates which player the actionSpace corresponds to.
ForwardModel now indicates the game type is for, with a virtual function getGameType().
GameState has new functions that indicate if a playerID can play in the state, and another one with the list of players that can play in this state.
GameState.currentPlayer is now -1 for RTS games.