NO_PLAYER_ID in GameRenderer set to -2 (before -1)
GameState.applyFogOfWar was particularly inefficient as it was creating a new vector with the entities of a given player. Now it runs over entities checking their ID and ignoring them if ownerID != playerID. BUT, if playerID==-1, this doesn't work, as -1 was used both for the observer's view from the GUI and the neutral player.